20240502 walk in the park
Problem Description
A robot dog is trying to take a walk in a park shaped like a rectangular grid, with 'O' indicating the passing path and 'X' indicating obstacles. The walk is carried out according to the commands pre-entered into the robot dog, and the commands are given in the following format.

["Direction distance", "Direction distance" … ]
For example, "E 5" means the robot dog has moved 5 spaces east from its current location. Before carrying out a command, the robot dog first checks two things:

Make sure you leave the park when traveling in a given direction.
Check whether any obstacles are encountered while moving in the given direction.
If either of the two above applies, the robot dog ignores the command and performs the next command.
If the horizontal length of the park is W and the vertical length is H, the coordinates of the upper left corner of the park are (0, 0) and the coordinates of the lower right corner are (H - 1, W - 1).

image

When a string array representing the park parkand a string array containing commands to be performed by the robot dog routesare given as parameters, the robot dog executes all commands and returns the location in the array in the order of [vertical coordinates, horizontal coordinates]. Please complete the solution function.

Restrictions
3 ≤ parklength ≤ 50
3 ≤ park[i]length ≤ 50
park[i]consists of the following characters and only one starting point is given.
S: starting point
O: movable passageway
X: Obstacle
parkis rectangular in shape.
1 ≤ routeslength ≤ 50
routesEach element in represents a command to be performed by the robot dog.
The robot dog routesperforms commands in order starting from the first element of .
routesThe elements have the same structure as "op n", where op means the direction to move and n means the number of cells to move.
op consists of one of the following four things:
N: Move north the given number of spaces.
S: Move south the given number of spaces.
W: Move west the given number of spaces.
E: Move east the given number of spaces.
1 ≤ n ≤ 9
Input/output example
park	routes	result
["SOO","OOO","OOO"]	["E 2","S 2","W 1"]	[2,1]
["SOO","OXX","OOO"]	["E 2","S 2","W 1"]	[0,1]
["OSO","OOO","OXO","OOO"]	["E 2","S 3","W 1"]	[0,0]
Input/Output Example Description
Input/Output Example #1

If you follow the entered command and move 2 spaces to the east, 2 spaces to the south, and 1 space to the west, the result becomes [0,0] -> [0,2] -> [2,2] -> [2,1].

Input/Output Example #2

According to the entered command, you must move 2 spaces to the east, 2 spaces to the south, and 1 space to the west, but when moving 2 spaces to the south, you pass through a space with an obstacle, so only commands other than that command are followed. The result is [0,0] -> [0,2] -> [0,1].

Input/Output Example #3

The first command entered will leave the park, and the second entered command will also pass the obstacle, so only the third command will be followed, excluding the two inputs, so the result is as follows. [0,1] -> [0,0]