20240503 덧칠하기
https://school.programmers.co.kr/learn/courses/30/lessons/161989
Problem Description
A school nhas a meter-long painted wall. Tape is often put on the wall to promote clubs and societies, or to post posters for company recruitment, but then removed when demolished, and the paint often peels off in the process. The peeling paint on the walls became unsightly, so the school decided to paint the walls again.

Instead of applying a fresh coat of paint to an entire large wall, try to save money by dividing it into sections and repainting just a few areas. To achieve this, the wall was divided into sections one meter long , and each section was numbered nfrom left to right, starting from 1 to 1 . nThen I decided on the areas that needed to be repainted.

The length of the roller for applying paint to the wall is meters, and the rules for applying one coatm of paint to the wall with a roller are as follows:

The roller must not move away from the wall.
Do not paint so that only part of the area is covered.
In other words, align the left and right ends of the roller with the boundary line of the area or the left and right ends of the wall, then move the roller up and down to paint the wall. The roller is removed from the wall after the currently painted areas have been completely painted, which defines the wall as having been painted once .

You can apply multiple coats of paint to an area, or to areas that do not need to be repainted, but any area you decide to repaint should be painted at least once. Just as you decide which areas to repaint to save on a budget, you also want to minimize the number of times you paint with a roller.

Write a solution function that returns the minimum number of times to paint with a roller when given as parameters an integer n, mand an array of integers containing the numbers of the areas to be repainted .section

Restrictions
1 ≤ m≤ n≤ 100,000
1 ≤ sectionlength ≤n
1 ≤ sectionelement of ≤n
sectionThe element of is the number of the area that needs to be repainted.
sectionThe same element does not appear more than once.
sectionThe elements are sorted in ascending order.
Input/output example
n	m	section	result
8	4	[2, 3, 6]	2
5	4	[1, 3]	One
4	One	[1, 2, 3, 4]	4
Input/Output Example Description
Input/Output Example #1

Example #1 requires repainting areas 2, 3, and 6. The roller is 4 meters long so you can cover four consecutive areas in one coat. If you paint areas 3, 4, 5, and 6 first, only area 2 will remain to be painted. If you paint areas 1, 2, 3, and 4, all areas that need to be repainted will be left in 2 coats. You can paint it.

img1.png

There is no way to apply paint to areas 2, 3, and 6 in less than 2 coats. Therefore, it returns 2, which is the minimum number of times.

Input/Output Example #2

Example 2 requires repainting areas 1 and 3. The roller is 4 meters long so you can paint four consecutive zones in one coat, and if you paint zones 1, 2, 3, and 4, you can paint zones 1 and 3 all in one coat.

img2.png

Therefore, it returns 1, which is the minimum number of times.

Input/Output Example #3

Example #3 requires all areas to be painted. The roller is 1 meter long so you can only paint one area at a time. Since there are 4 zones, the minimum number of times is 4, with each zone only being painted once.

img3.png

Therefore, it returns 4.

※ Notice - On March 20, 2023, the difficulty level of the question was Lv. 2 → Lv. Changed to 1.
※ Notice - A typo in the question on August 17, 2023 has been corrected.