20240503 대충 만든 자판
https://school.programmers.co.kr/learn/courses/30/lessons/160586
Problem Description
Unlike computer keyboard keyboards, cell phone keyboards can have multiple characters assigned to one key. If multiple letters are assigned to a key, pressing the same key in quick succession changes the letters in the order in which they are assigned.

For example, if the letters "A", "B", and "C" are assigned to key 1 in that order, pressing key 1 once will give you "A", pressing it twice will get "B", and pressing it three times will get "C". And so on.

There are cell phone keyboards made haphazardly using the same rules. This mobile phone keyboard can have from 1 to a maximum of 100 keys, and the characters entered when certain keys are pressed are arranged randomly. Also, there are cases where the same letter is assigned multiple times across the entire keyboard, and there are cases where the same letter is assigned multiple times to a single key. In some cases, it is not even allocated at all. Therefore, some strings may not be able to be written.

When writing a specific string using this mobile phone keyboard, we want to find out how many keys must be pressed at least several times to write that string.

Given a string array containing the characters assigned in order starting from key 1 keymapand a string array containing the strings to be input targets, a solution function is provided that returns the minimum number of times the key must be pressed to write each string in an array in order. Please complete it.

However, if the target string cannot be created, -1 is stored.

Restrictions
1 ≤ keymaplength ≤ 100
1 ≤ keymapelement length ≤ 100
keymap[i]means the characters that change in order when you press the i + 1 key.
For example, keymap[0]if = "ABACD", pressing key 1 once will result in A, pressing it twice will result in B, and pressing it three times will result in A.
keymapThe lengths of the elements may be different.
keymapThe elements of consist of only uppercase letters of the alphabet.
1 ≤ targetslength ≤ 100
1 ≤ targetselement length ≤ 100
targetsThe elements of consist of only uppercase letters of the alphabet.
Input/output example
keymap	targets	result
["ABACD", "BCEFD"]	["ABCD","AABB"]	[9, 4]
["AA"]	["B"]	[-One]
["AGZ", "BSSS"]	["ASA","BGZ"]	[4, 6]
Input/Output Example Description
Input/Output Example #1

For "ABCD",
Press key 1 once → A
Press key 2 once → B
Press 2 key twice → C
Press 1 key five times → D
So we store the total, 9, in the first index.
For "AABB",
Press key 1 once → A
Press key 1 once → A
Press key 2 once → B
Press key 2 once → B
So we store the total, 4, in the second index.
As a result, [9,4] is returned.
Input/Output Example #2

For "B", we store -1 in the first index because 'B' does not exist anywhere.
As a result, [-1] is returned.
Input/Output Example #3

For "ASA",
Press key 1 once → A
Press key 2 twice → S
Press key 1 once → A
So we store the total of 4 in the first index.
For "BGZ",
Press key 2 once → B
Press 1 key twice → G
Press number 1 three times → Z
So we store the total, 6, in the second index.
As a result, [4, 6] is returned.