20240504 개인정보 수집 유효기간
https://school.programmers.co.kr/learn/courses/30/lessons/150370
Problem Description
There is npersonal information classified as 1~ collected with the customer's consent to the terms and conditions . nThere are several types of terms and conditions, and each term has a set expiration date for storing personal information. You know under what terms each piece of personal information is collected. Collected personal information can only be stored until the expiration date, and if the expiration date has passed, it must be destroyed.

For example, if the validity period of Terms and Conditions A is 12 months and personal information collected on January 5, 2021 was collected under Terms and Conditions A, the personal information can be stored until January 4, 2022 and January 5, 2022. This is personal information that must be destroyed first.
You are trying to obtain personal information numbers that need to be destroyed as of today.

All months are assumed to have 28 days.

Below 2022.05.19is an example where today's date is:

Terms and Conditions Type	expiration period
A	6 months
B	12 months
C	3 months
number	Personal information collection date	Terms and Conditions Type
One	2021.05.02	A
2	2021.07.01	B
3	2022.02.19	C
4	2022.02.20	C
The first personal information can be stored until November 1, 2021 in accordance with Terms A, and as it has expired, it is personal information that must be destroyed.
The second personal information can be stored until June 28, 2022 according to Terms and Conditions B. Since the expiration date has not passed, it can still be stored.
Third, personal information can be stored until May 18, 2022 according to Terms and Conditions C, and as it has expired, it is personal information that must be destroyed.
Fourth, personal information can be stored until May 19, 2022 in accordance with Terms C, and since the expiration date has not passed, it can still be stored.
Therefore, the personal information number to be destroyed is [1, 3].

A string indicating today's date today, a one-dimensional string array containing the validity period of the terms and conditions, termsand a one-dimensional string array containing collected personal information privaciesare given as parameters. At this time, complete the solution function to return the number of personal information that needs to be destroyed in ascending order in a one-dimensional integer array.

Restrictions
todayrepresents today's date in the form " YYYY. MM. ".DD
1 ≤ termslength ≤ 20
termsThe element of is a string of the form " " 약관 종류 유효기간separated by a space.약관 종류유효기간
약관 종류is one uppercase letter of the alphabet, and Athere are no duplicates in the array .Zterms약관 종류
유효기간is an integer representing the number of months for which personal information can be stored, and must be between 1 and 100.
1 ≤ privacieslength ≤ 100
privacies[i]i+1Indicates the date of collection of personal information and the type of terms and conditions .
privaciesThe element of is a string of " " 날짜 약관 종류form separated by a space.날짜약관 종류
날짜indicates the date on which personal information in the form of " YYYY. MM. " was collected, and only previous dates are given.DDtoday
privaciesof is 약관 종류always given only those termsthat appear in .약관 종류
todayprivaciesAppearing in and , 날짜represents YYYYthe year, represents MMthe month, DDand represents the day, and .is separated by a single dot ( ).
2000 ≤ YYYY≤ 2022
1 ≤ MM≤ 12
MMIf this is a single digit number, a leading 0 is added.
1 ≤ DD≤ 28
DDIf is a single digit, a leading 0 is added.
Only inputs that contain at least one piece of personal information that needs to be destroyed are given.
Input/output example
today	terms	privacies	result
"2022.05.19"	["A 6", "B 12", "C 3"]	["2021.05.02 A", "2021.07.01 B", "2022.02.19 C", "2022.02.20 C"]	[1, 3]
"2020.01.01"	["Z 3", "D 5"]	["2019.01.01 D", "2019.11.15 Z", "2019.08.02 D", "2019.07.01 D", "2018.12.28 Z"]	[1, 4, 5]
Input/Output Example Description
Input/Output Example #1

It's the same as the example problem.
Input/Output Example #2

Terms and Conditions Type	expiration period
Z	3 months
D	5 months
number	Personal information collection date	Terms and Conditions Type
One	2019.01.01	D
2	2019.11.15	Z
3	2019.08.02	D
4	2019.07.01	D
5	2018.12.28	Z
Today's date is January 1, 2020.

The first personal information can be stored until May 28, 2019 in accordance with the D terms and conditions, and as it has expired, it is personal information that must be destroyed.
The second personal information can be stored until February 14, 2020 according to the Z Terms and Conditions. Since it has not expired, it can still be stored.
Third, personal information can be stored until January 1, 2020 in accordance with Terms D, and can still be stored because the expiration date has not passed.
Fourth, personal information can be stored until November 28, 2019 in accordance with the D Terms and Conditions, and as it has expired, it is personal information that must be destroyed.
Fifth, personal information can be stored until March 27, 2019 in accordance with the Z Terms and Conditions, and as it has expired, it is personal information that must be destroyed.