20240504 크기가 작은 부분 문자열
https://school.programmers.co.kr/learn/courses/30/lessons/147355

Problem Description
Given a string of numbers tand , complete a function solution that returns the number of times the number represented by this substring is less than or equal to the number represented by among substrings with the same length as .ptpp

For example, if t="3141592" and ="271" , the substrings of length 3 are 314, 141, 415, 159, 592. Among the numbers represented by this string, the two numbers less than or equal to 271 are 141 and 159.pt

Restrictions
1 ≤ plength ≤ 18
pLength of ≤ tLength of ≤ 10,000
tand pare strings consisting only of numbers and do not start with 0.
Input/output example
t	p	result
"3141592"	"271"	2
"500220839878"	"7"	8
"10203"	"15"	3
Input/Output Example Description
Input/Output Example #1
Same as the main text.

pSince the length of input/output example #2 is 1 t, the substrings of are "5", "0", 0", "2", "2", "0", "8", "3", "9", "8" ", "7", and "8", and numbers less than or equal to 7 are "5", "0", "0", "2", "2", "0", "3", and "7". There are 8.

pSince the length of input/output example #3 is 2 t, the substrings of are "10", "02", "20", and "03", of which numbers less than or equal to 15 are "10", "02", and "03". There are three. Note that "02" and "03" correspond to 2 and 3 respectively