20240505 명예의 전당 (1)
https://school.programmers.co.kr/learn/courses/30/lessons/138477

Problem Description
In the TV program "Hall of Fame," one singer sings a song every day, and points are awarded to the singer based on the number of text votes from viewers. If the score of a singer who appears every day is within the top kth scores of all singers who have appeared so far, that singer's score is commemorated by being added to a list called the Hall of Fame. In other words, from the beginning of the program until k days, the scores of all participating singers will be entered into the Hall of Fame. From k days onwards, if the score of the appearing singer is higher than the score of the kth ranked singer in the existing Hall of Fame list, the featured singer's score will be promoted to the Hall of Fame, and the existing kth ranked singer will be lowered from the Hall of Fame. It will come.

The program announces the lowest scores of the “Hall of Fame” each day. For example, kif = 3 and the singer's score over 7 days is [10, 100, 20, 150, 1, 100, 200], the score announced at the Hall of Fame is as shown below. [10, 10, 10, 20, 20, 100, 100].

Given the number of scores in the Hall of Fame list kand the scores of singers who appeared from the 1st to the last day , please complete the solution function that returns the lowest score of the Hall of Fame announced each day.score

Restrictions
3 ≤ k≤ 100
7 ≤ scorelength ≤ 1,000
0 ≤ score[i]≤ 2,000
Input/output example
k	score	result
3	[10, 100, 20, 150, 1, 100, 200]	[10, 10, 10, 20, 20, 100, 100]
4	[0, 300, 40, 300, 20, 70, 150, 50, 500, 1000]	[0, 0, 0, 0, 20, 40, 70, 70, 150, 300]
Input/Output Example Description
Input/Output Example #1

This is an example of the problem.
Input/Output Example #2

As shown below, it returns [0, 0, 0, 0, 20, 40, 70, 70, 150, 300].