20240505 기사단원의 무기
https://school.programmers.co.kr/learn/courses/30/lessons/136798

Problem Description
numberEach knight of the Order of Numbers is assigned a number from 1 to 2 . The knights are trying to purchase weapons at the weapon shop.

Each knight attempts to purchase a weapon with an attack power equal to the number of divisors of his or her knight number. However, a limit on attack power is set by an agreement with a neighboring country, and knights who must purchase weapons with attack power greater than the limit must purchase weapons with attack power determined by the agreement agency.

For example, a templar designated as number 15 purchases a weapon with an attack power of 4 because 15 has 4 divisors: 1, 3, 5, and 15. If the attack power limit set by an agreement with a neighboring country is 3, and the weapon used by the knight who exceeds the limit has an attack power of 2, the knight designated as number 15 purchases a weapon with an attack power of 2 at the weapon store. When making a weapon, 1 kg of iron is required per 1 attack power of the weapon. So, I'm trying to calculate in advance the weight of iron needed to make all the weapons at the weapon shop.

Given an integer representing the number of knights number, an integer representing the attack power limit determined by agreements with neighboring countries, limitand an integer representing the attack power of weapons used by knights exceeding the limit , the owner of the weapon shop must make all the weapons. powerComplete the solution function that returns the required weight of iron.

Restrictions
1 ≤ number≤ 100,000
2 ≤ limit≤ 100
1 ≤ power≤limit
Input/output example
number	limit	power	result
5	3	2	10
10	3	2	21
Input/Output Example Description
Input/Output Example #1

The number of divisors from 1 to 5 in order is [1, 2, 2, 3, 2]. Since none of them exceed the attack power limit of 3, the weight of iron required is the sum of those numbers, which is 10. Therefore, it returns 10.

Input/Output Example #2

The number of divisors from 1 to 10 in order is [1, 2, 2, 3, 2, 4, 2, 4, 3, 4]. Since the attack power limit is 3, knights 6, 8, and 10 purchase weapons with an attack power of 2. Therefore, it returns 21, which is the sum of those numbers.