20240505 과일 장수
https://school.programmers.co.kr/learn/courses/30/lessons/135808

Problem Description
A fruit seller is packing boxes of apples. Apples are classified into scores from 1 to k depending on their condition, with k being the best quality apple and 1 being the worst quality apple. The price of a box of apples is determined as follows:

Each box is packed with m apples.
If the lowest score of apples in the box is p (1 ≤ p ≤ k), then the price of a box of apples is p * m.
We want to calculate the maximum profit that a fruit seller can make if he sells as many apples as possible (apples are only sold in boxes, and any excess apples are discarded).

For example, k= 3, m= 4, if the score of 7 apples is [1, 2, 3, 1, 2, 3, 1], then 1 box of apples consists of [2, 3, 2, 3] as follows: You can make and sell them for maximum profit.

(lowest apple score) x (number of apples in a box) x (number of boxes) = 2 x 4 x 1 = 8
Given the maximum score of apples k, the number of apples in one box m, and the scores of apples , please complete the solution function that returns the maximum profit that the fruit seller can obtain.score

Restrictions
3 ≤ k≤ 9
3 ≤ m≤ 10
7 ≤ scorelength ≤ 1,000,000
1 ≤ score[i]≤ k
If no profit is generated, please return 0.
Input/output example
k	m	score	result
3	4	[1, 2, 3, 1, 2, 3, 1]	8
4	3	[4, 1, 2, 2, 4, 4, 4, 4, 1, 2, 4, 2]	33
Input/Output Example Description
Input/Output Example #1

This is an example of the problem.
Input/Output Example #2

You can make maximum profit by packing crates of apples and selling them all as follows:
apple box	price
[1, 1, 2]	1 x 3 = 3
[2, 2, 2]	2 x 3 = 6
[4, 4, 4]	4 x 3 = 12
[4, 4, 4]	4 x 3 = 12
Therefore, it returns (1 x 3 x 1) + (2 x 3 x 1) + (4 x 3 x 2) = 33.