20240507 콜라 문제
https://school.programmers.co.kr/learn/courses/30/lessons/132267

Problem Description
There is a coke problem that was popular a long time ago. The fingerprint of the Coke problem is as follows:

Don't tell anyone the answer.

There is a supermarket that gives you a bottle of Coke if you bring two empty Coke bottles. If I bring 20 empty bottles, how many bottles will I get?

However, if you have less than 2 empty bottles, you cannot receive Coke.

While solving the problem, Sangbin found the perfect answer to the cola problem. Sangbin’s solution is shown in the picture below. First, take 20 empty bottles of Coke and receive 10 bottles. After drinking all 10 bottles you received, take them to receive 5 bottles. After drinking all 4 of the 5 bottles, take them back to receive 2 bottles, and after drinking all 2 bottles, take them back to receive 1 bottle. When you receive bottles 1 and 5, you can receive another bottle if you drink the remaining bottle and take it back. In this case, Sangbin can receive a total of 10 + 5 + 2 + 1 + 1 = 19 bottles of cola.

Sangbin, who was working hard to solve the problem, thought of the generalized coke problem. This problem is to calculate how many bottles you can get if ayou bring empty bottles bto a supermarket that gives you a Coke bottle . nAs with the existing Coke problem, aif you have less than 1 empty bottle, you cannot receive additional empty bottles. Sangbin thought hard, but he couldn't find the answer to the generalized coke problem. Please help Sangbin and create a program that solves the general coke problem.

The number of bottles that must be given to the mart to receive cola a, the number of cola bottles that the mart gives if you bring an empty bottle a b, and the number of empty bottles that Sangbin has nare given as parameters. Please write a solution function to return the number of bottles of cola that Sangbin can receive.

Restrictions
1 ≤ b< a≤ n≤ 1,000,000
The correct answer is always given within the int range.
Input/output example
a	b	n	result
2	One	20	19
3	One	20	9
Input/Output Example Description
Input/Output Example #1

This is an example described in the text.
Input/Output Example #2

Take 18 out of 20 empty bottles to the supermarket and receive 6 bottles of Coke. At this time, the number of Coke bottles Sangbin has is 8 (20 – 18 + 6 = 8).
Take 6 out of 8 empty bottles to the supermarket and receive 2 bottles of Coke. At this time, the number of Coke bottles Sangbin has is 4 (8 – 6 + 2 = 4).
Take 3 out of 4 empty bottles to the supermarket and receive 1 bottle of Coke. At this time, the number of Coke bottles Sangbin has is 2 (4 – 3 + 1 = 2).
During the three exchanges, Sangbin received 9 (6 + 2 + 1 = 9) bottles of Coke.