20240507 삼총사
https://school.programmers.co.kr/learn/courses/30/lessons/131705

Problem Description
Each student attending a Korean middle school has an integer number. If the integer numbers of the three students at this school add up to 0, the three students are called the Three Musketeers. For example, if there are five students, and each integer number is -2, 3, 0, 2, and -5 in that order, the integer numbers of the first, third, and fourth students add up to 0, so the three students are It's the Three Musketeers. Additionally, the integer numbers of the second, fourth, and fifth students add up to 0, so the three students are also a trio. Therefore, in this case, Korean middle schools can create the Three Musketeers in two ways.

When an integer array representing the numbers of Korean middle school students numberis given as a parameter, complete the solution function to return the number of ways to create a trio among the students.

Restrictions
3 ≤ numberlength ≤ 13
-1,000 ≤ numberEach element of ≤ 1,000
Different students may have the same integer number.
Input/output example
number	result
[-2, 3, 0, 2, -5]	2
[-3, -2, -1, 0, 1, 2, 3]	5
[-1, 1, -1, 1]	0
Input/Output Example Description
Input/Output Example #1

It's the same as the example problem.
Input/Output Example #2

Students' integer number pairs (-3, 0, 3), (-2, 0, 2), (-1, 0, 1), (-2, -1, 3), (-3, 1, 2) Since this can be the Three Musketeers, we return 5.
Input/Output Example #3

There is no way it could be the Three Musketeers.