20240515 완주하지 못한 선수
https://school.programmers.co.kr/learn/courses/30/lessons/42576

Problem Description
Numerous marathon runners participated in the marathon. All but one athlete completed the marathon.

When given an array participant containing the names of athletes who participated in the marathon and an array completion containing the names of athletes who completed the marathon, write a solution function to return the names of athletes who did not complete the marathon.

Restrictions
The number of athletes participating in a marathon is between 1 and 100,000.
The length of completion is 1 less than the length of the participant.
The participant's name must consist of at least 1 and not more than 20 lowercase letters.
Among the participants, there may be people with the same name.
Input/output example
participant	completion	return
["leo", "kiki", "eden"]	["eden", "kiki"]	"leo"
["marina", "josipa", "nikola", "vinko", "filipa"]	["josipa", "filipa", "marina", "nikola"]	"vinko"
["mislav", "stanko", "mislav", "ana"]	["stanko", "ana", "mislav"]	"mislav"
Input/Output Example Description
Example #1
"leo" is in the participant list, but not in the finisher list, so he did not finish the race.

Example #2
“vinko” did not finish because he was on the participant list but not on the finishers list.

Example #3
"mislav" has two people on the participant list, but only one person on the finisher list, so one of them did not finish.

※ Notice - Test case was added on January 25, 2023.