20240517 이중우선순위큐
https://school.programmers.co.kr/learn/courses/30/lessons/42628

Problem Description
A double priority queue refers to a data structure that can perform the following operations.

command	Receiving Tower (Height)
I number	Inserts the given number into the queue.
D 1	Remove the maximum value from the queue.
D-1	Removes the minimum value from the queue.
When the operations to be performed by the double priority queue are given as parameters, implement the solution function to return [0,0] if the queue is empty after processing all operations, and [maximum value, minimum value] if it is not empty.

Restrictions
operations is an array of strings with length between 1 and 1,000,000.
The elements of operations represent the operations to be performed by the queue.
Elements are given in “command data” format. - In the operation of deleting the maximum/minimum value, if there is more than one maximum/minimum value, only one is deleted.
If an operation is given to delete data from an empty queue, that operation is ignored.
Input/output example
operations	return
["I 16", "I -5643", "D -1", "D 1", "D 1", "I 123", "D -1"]	[0,0]
["I -45", "I 653", "D 1", "I -642", "I 45", "I 97", "D 1", "D -1", "I 333"]	[333, -45]
Input/Output Example Description
Input/Output Example #1

Insert 16 and -5643.
Delete the minimum value. -5643 is deleted, leaving 16.
Delete the maximum value. 16 is deleted and the dual priority queue is empty.
Since the priority queue is empty, the max delete operation is ignored.
Insert 123.
Delete the minimum value. 123 is deleted and the dual priority queue is empty.
So it returns [0, 0].

Input/Output Example #2

After inserting -45 and 653, delete the maximum value (653). -45 left.
After inserting -642, 45, and 97, the maximum value (97) and minimum value (-642) are deleted. -45 and 45 remain.
Insert 333.
There are -45, 45, 333 left in the double priority queue, so it returns [333, -45].